creates pipeline abstraction to represent pfcp chain (pdr and actions)

+++ src/bpf/far_maps.h
@@ -9,7 +9,7 @@
 // The unique FAR that will be consumed in this program.
 struct bpf_map_def SEC("maps") m_far = {
     .type = BPF_MAP_TYPE_HASH,
-    .key_size = sizeof(u32),
+    .key_size = sizeof(u8),
     .value_size = sizeof(pfcp_far_t_),
     .max_entries = 1,
 };
